LRDOCS-8138 Run SF
