+++ docs/dxp/7.x/en/process-automation/forms/developer-guide/writing-a-form-storage-adapter/liferay-r2f1.zip/r2f1-impl/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Acme R2F1 DDM Form Storage Adapter 
+Bundle-Name: Acme R2F1 DDM Form Storage Adapter
 Bundle-SymbolicName: com.liferay.learn.r2f1.impl
 Bundle-Version: 1.0.0
 Provide-Capability:\
@@ -6,7 +6,7 @@ Provide-Capability:\
 		resource.bundle.base.name="content.Language",\
 	liferay.resource.bundle;\
 		bundle.symbolic.name=com.liferay.dynamic.data.mapping.lang;\
-		resource.bundle.aggregate:String="(bundle.symbolic.name=com.liferay.dynamic.data.mapping.lang),(bundle.symbolic.name=com.acme.r2f1.impl)";\
+		resource.bundle.aggregate:String="(bundle.symbolic.name=com.acme.r2f1.impl),(bundle.symbolic.name=com.liferay.dynamic.data.mapping.lang)";\
 		resource.bundle.base.name="content.Language";\
 		service.ranking:Long="2";\
 		servlet.context.name=dynamic-data-mapping-form-web
