+++ docs/dxp/7.x/en/process-automation/forms/developer-guide/writing-a-form-storage-adapter/liferay-r2f1.zip/r2f1-impl/src/main/java/com/acme/r2f1/internal/data/mapping/storage/DDMFileSystemStorageAdapter.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.acme.r2f1.impl;
+package com.acme.r2f1.internal.data.mapping.storage;
 
 import com.liferay.dynamic.data.mapping.exception.StorageException;
 import com.liferay.dynamic.data.mapping.io.DDMFormValuesDeserializer;
