+++ docs/dxp/7.x/en/site-building/developer-guide/developing-page-fragments/creating-a-contributed-fragment-collection/liferay-l3m9.zip/l3m9-impl/bnd.bnd
