+++ docs/dxp/7.x/en/site-building/developer-guide/developing-page-fragments/creating-a-contributed-fragment-collection/liferay-l3m9.zip/l3m9-impl/marketing-jumbotron/index.html
@@ -1,17 +1,21 @@
 <div class="marketing-jumbotron-fragment-01">
-  <div class="jumbotron">
-    <lfr-editable id="01-jumbotron-title" type="rich-text">
-      <h1 class="display-4">Editable Jumbotron headline</h1>
-    </lfr-editable>
-    <lfr-editable id="02-jumbotron-call" type="rich-text">
-      <p class="lead">Edit this text to call extra attention to featured content or information.</p>
-    </lfr-editable>
-    <hr class="my-4">
-    <lfr-editable id="03-jumbotron-more-info" type="rich-text">
-      <p>Edit this text to provide more information.</p>
-    </lfr-editable>
-    <lfr-editable id="04-jumbotron-link" type="link">
-      <a class="btn btn-primary btn-lg" href="#" role="button">Editable link</a>
-    </lfr-editable>
-  </div>
+<div class="jumbotron">
+	<lfr-editable id="01-jumbotron-title" type="rich-text">
+	  <h1 class="display-4">Editable Jumbotron headline</h1>
+	</lfr-editable>
+
+	<lfr-editable id="02-jumbotron-call" type="rich-text">
+	  <p class="lead">Edit this text to call extra attention to featured content or information.</p>
+	</lfr-editable>
+
+	<hr class="my-4" />
+
+	<lfr-editable id="03-jumbotron-more-info" type="rich-text">
+	  <p>Edit this text to provide more information.</p>
+	</lfr-editable>
+
+	<lfr-editable id="04-jumbotron-link" type="link">
+	  <a class="btn btn-primary btn-lg" href="#" role="button">Editable link</a>
+	</lfr-editable>
+</div>
 </div>
