--- docs/dxp/7.x/en/site-building/developer-guide/developing-page-fragments/creating-a-contributed-fragment-collection/liferay-l3m9.zip/l3m9-impl/src/main/java/com/liferay/learn/fragments/FragmentsCollectionContributor.java
@@ -1,35 +1,0 @@
-package com.liferay.learn.fragments;
-
-import javax.servlet.ServletContext;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-import com.liferay.fragment.contributor.BaseFragmentCollectionContributor;
-import com.liferay.fragment.contributor.FragmentCollectionContributor;
-
-/**
- * @author liferay
- */
-@Component(
-	service = FragmentCollectionContributor.class
-)
-public class FragmentsCollectionContributor extends BaseFragmentCollectionContributor {
-
-	@Override
-	public String getFragmentCollectionKey() {
-	    return "Marketing Collection";
-	}
-	
-	@Override
-	public ServletContext getServletContext() {
-	    return _servletContext;
-	}
-	
-	@Reference(
-    target = "(osgi.web.symbolicname=com.liferay.learn.fragments)"
-	)
-	
-	private ServletContext _servletContext;
-
-}
