+++ docs/dxp/7.x/en/site-building/developer-guide/developing-page-fragments/creating-a-contributed-fragment-collection/liferay-l3m9.zip/l3m9-impl/src/main/resources/com/acme/l3m9/fragments/dependencies/collection.json
@@ -1,6 +1,6 @@
 {
-    "fragments": [
-        "marketing-card"
-    ],
-    "name": "Marketing Collection"
+	"fragments": [
+		"marketing-card"
+	],
+	"name": "Marketing Collection"
 }
