+++ docs/dxp/7.x/en/site-building/developer-guide/developing-page-fragments/creating-a-contributed-fragment-collection/liferay-l3m9.zip/l3m9-impl/src/main/resources/com/acme/l3m9/fragments/dependencies/marketingcard/index.html
@@ -1,18 +1,19 @@
 <div class="marketing-card-fragment-01">
-  <div class="card">
-    <lfr-editable id="01-card-image" type="image">
-      <img src="https://clayui.com/images/home/lexicon_symbol.svg" class="card-img-top" alt="Clay UI Logo">
-    </lfr-editable>
-    <div class="card-body">
-      <lfr-editable id="02-card-title" type="rich-text">
-        <h5 class="card-title">Editable Card title</h5>
-      </lfr-editable>
-      <lfr-editable id="03-card-text" type="rich-text">
-        <p class="card-text">Here is some editable text.</p>
-      </lfr-editable>
-      <lfr-editable id="04-card-link" type="link">
-        <a href="#" class="btn btn-primary">Editable link</a>
-      </lfr-editable>
-    </div>
-  </div>
+<div class="card">
+	<lfr-editable id="01-card-image" type="image">
+	<img alt="Clay UI Logo" class="card-img-top" src="https://clayui.com/images/home/lexicon_symbol.svg" />
+	</lfr-editable>
+
+	<div class="card-body">
+	<lfr-editable id="02-card-title" type="rich-text">
+		<h5 class="card-title">Editable Card title</h5>
+	  </lfr-editable>
+	<lfr-editable id="03-card-text" type="rich-text">
+		<p class="card-text">Here is some editable text.</p>
+	  </lfr-editable>
+	<lfr-editable id="04-card-link" type="link">
+		<a href="#" class="btn btn-primary">Editable link</a>
+	  </lfr-editable>
+	</div>
+</div>
 </div>
