+++ docs/dxp/7.x/en/using-search/developer-guide/writing-a-similar-results-contributor/liferay-r1s1.zip/r1s1-impl/build.gradle
